Authentication utils
I created some helper functions in order to make authentication easier. getUserData
gets the information from localStorage and returns an object with the data.

handleAuthentication gets the data, checks if there is an older token and
finally sets the authenticated state according with the expiry date.

saveAuthStorageHandler will get the data and set the items in localStorage.
On the other hand, logoutStorageHandler does the opposite, removes the data from localStorage