Managing user authentication state
I created an authSlice that handles the reducer and actions creators. I needed
to pass an object with the slice name, initial data (from localStorage) and the reducers
object that defines how to update the state when an action is dispatched.
Then, I got the actions from the slice and exports them along with the reducer.

The next step is creating a store, passing the reducer object and then,
I specified the auth slice from the whole state and assign it to the previous reducer.

Finally, I connected my app with the Redux central store passing it to the Provide component as props.