Login and Register pages
The Login component renders a Formik form that handles validation and form state automatically.
When the user submits the form, it calls an eventHandler that makes an axios POST
request, sending the form data to the backend and dispatching and action with
the data returned by the backend.
The reducers gets the payload, saves the data on localStorage and updates the main state.

Register.js is the same as the Login component, but it has an aditional form field and
it also sends the data to a different endpoint.