Creating API slice
According with Redux Toolkit, I can create an Api slicer, so that it creates
actions, reducers and hooks automatically and makes API request easier.

The first step is create and api object, pass some reducer path to it,
in order to give the reducer a unique name. Then, a baseQuery that acts kind
of similar to the JS fetch API, passing the baseUrl to it.

Finally, I had to define a set of endpoints. For now, I just have a getCoins
endpoint that gets a query with the url and the headers.