Cryptocurrencies.js Logic
This component gets a simplified props that checks whether I want to display
few crypto cards or the whole response. If that's the case, I pass the 99
limit to my hook created by the api slice and fetch the data. Finally, I map
through the cryptos data and renders each card component.